+++ src/auth_groups/Controllers/AuthController.php
@@ -11,6 +11,7 @@ use AuthGroups\Services\DeviceTokenService;
 use AuthGroups\Services\EmailService;
 use AuthGroups\Services\LogService;
 use AuthGroups\Services\UserSessionService;
+use AuthGroups\Services\SubscriptionService;
 use AuthGroups\Utils\Response;
 use AuthGroups\Utils\Validator;
 use AuthGroups\Middleware\LoggingMiddleware;
@@ -420,6 +421,7 @@ class AuthController
                 'email' => $userData['email'],
                 'role'  => $userData['role'],
             ],
+            'subscriptions' => SubscriptionService::getAllStatuses((int) $userData['id']),
         ];
 
         if ($deviceToken !== null) {